use lunr to search related posts
